+++ apps/java/test_app/src/main/java/com/test_app/Primes.java
@@ -9,7 +9,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 class Primes {
-    static boolean isPrime(long num) {
+    static boolean isPrime(int num) {
         if (num <= 1) return false;
         if (num == 2) return true;
         if (num % 2 == 0) return false; // the only even prime is 2
@@ -17,27 +17,27 @@ class Primes {
         // Trial division (https://en.wikipedia.org/wiki/Prime_number#Trial_division)
         var limit = Math.sqrt(num);
 
-        for (long i=3; i<=limit; i++) {
+        for (int i=3; i<=limit; i++) {
             if (num % i == 0) return false;
         }
 
         return true;
-    } 
+    }
 
-    static List<Long> getPrimesLessThan(long limit) {
-        var result = new ArrayList<Long>();
+    static List<Integer> getPrimesLessThan(int limit) {
+        var result = new ArrayList<Integer>();
 
-        for (long i=2; i<limit; i++) {
+        for (int i=2; i<limit; i++) {
             if (Primes.isPrime(i)) result.add(i);
         }
 
         return result;
     }
 
-    static long countPrimesLessThan(long limit) {
-        long result = 0;
+    static int countPrimesLessThan(int limit) {
+        int result = 0;
 
-        for (long i=2; i<limit; i++) {
+        for (int i=2; i<limit; i++) {
             if (Primes.isPrime(i)) result++;
         }
